requirements.txt
streamlit
fal_client
python-dotenv
requests
Pillow